refactor: Queue 분리
