--- src/main/java/kr/hhplus/be/server/concert/domain/queue/QueueStatus.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.concert.domain.queue;
-
-public enum QueueStatus {
-	WAITING,
-	ACTIVE
-}
