+++ src/main/java/kr/hhplus/be/server/queue/adapter/in/web/response/QueueTokenResponse.java
@@ -1,11 +1,11 @@
-package kr.hhplus.be.server.reservation.adapter.in.web.queue.response;
+package kr.hhplus.be.server.queue.adapter.in.web.response;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.hhplus.be.server.concert.domain.queue.QueueStatus;
-import kr.hhplus.be.server.concert.domain.queue.QueueToken;
+import kr.hhplus.be.server.queue.domain.QueueStatus;
+import kr.hhplus.be.server.queue.domain.QueueToken;
 import lombok.Builder;
 
 @Builder
