+++ src/main/java/kr/hhplus/be/server/queue/domain/QueueStatus.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.queue.domain;
+
+public enum QueueStatus {
+	WAITING,
+	ACTIVE
+}
