+++ src/main/java/kr/hhplus/be/server/queue/port/in/GetActiveQueueTokenUseCase.java
@@ -1,6 +1,6 @@
-package kr.hhplus.be.server.concert.port.in.queue;
+package kr.hhplus.be.server.queue.port.in;
 
-import kr.hhplus.be.server.concert.domain.queue.QueueToken;
+import kr.hhplus.be.server.queue.domain.QueueToken;
 import kr.hhplus.be.server.common.exception.CustomException;
